refactor(docs-builder): Clean up & move paths
